requirements.txt
arrow==0.13.1
bibtexparser==1.1.0
blessed==1.15.0
chardet==3.0.4
Django==2.2.18
django-crispy-forms==1.7.2
django-environ==0.4.5
django-model-utils==3.1.2
django-picklefield==2.0
django-q==1.0.1
django-settings-export==1.2.1
django-simple-history==2.7.2
django-split-settings==1.0.1
django-sslserver==0.20
django-su==0.8.0
doi2bib==0.3.0
factory-boy==2.12.0
Faker==3.0.0
future==0.17.1
humanize==0.5.1
idna==2.8
pyparsing==2.3.1
python-dateutil==2.8.0
python-keystoneclient
python-memcached==1.59
pytz==2018.9
redis==3.2.1
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.3
urllib3==1.24.2
wcwidth==0.1.7
